added some fixes

+++ Practice/PracticeSix/file_handling/copy_delete_files.py
@@ -7,11 +7,11 @@ copy_path = os.path.join(BASE_DIR, "sample_copy.txt")
 
 shutil.copy(sample_path, copy_path)
 
-print("\nФайл успешно скопирован в sample_copy.txt")
+print("\nФайл скопирован в sample_copy.txt")
 
 file_to_delete = copy_path
-
-if os.path.exists(file_to_delete):
+check = bool(input("Delete file?"))
+if (os.path.exists(file_to_delete) and check == True):
     os.remove(file_to_delete)
     print(f"\nФайл {file_to_delete} удален.")
 else:
